Fix PI DMA test (#66)
The current version of the test runs a PI DMA from PI address B0xx_xxxx
(erroneously using a CPU virtual address in the uncached segment as PI
address).

This happens to work on Everdrive X7 because Everdrive possibly reproduces
the simplified address decoding logic used by game cartridges, where
only bits A27 and A28 are taken into account. So in a way the test would
pass if it was flashed into a real mask ROM mounted on a PCB connected
to the PI bus via the simplified logic.

64drive and Summercart64 instead do not emulate this kind of simplified
address mapping: ROMs are only mapped at 10xx_xxxx.

The goal of this test is to make sure emulators correctly emulate the
PI bus. What the N64 RCP PI does is to forward the full 32 bit
address from the register to the bus. It is probably better to induce
emulators into this kind of behavior (as it's easier to add masking
on top, at the cartridge level), rather than making them confused
on CPU virtual addresses vs PI addresses.